Move FileAttributesEtl Lambda into subfolder
